Move docker types to v1 API
